Optimise Problem 24
Reduce runtime from ~8 seconds to < 1.5 seconds.

Instead of generating a list of all permutations out-of-order and then
sorting, generate a stream of permutations in order and stop as soon as
the `n`th is reached. Also, replace all instances of `List<Integer>`
with `int[]`.

TODO: There is still room for further optimisation. With the right
maths, it should be possible to skip straight to the `n`th permutation.

- Problem 24